# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "推送通知基础服务，支持多平台推送和用户令牌管理",
        "title": "推送基础服务 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.idchat.io",
    "basePath": "/push-base",
    "paths": {
        "/v1/push/add_blocked_chat": {
            "post": {
                "description": "为用户添加屏蔽某个群聊或私聊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push API"
                ],
                "summary": "添加屏蔽聊天",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddBlockedChatReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/v1/push/get_user_blocked_chats": {
            "get": {
                "description": "根据用户 metaId 获取该用户屏蔽的聊天列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push API"
                ],
                "summary": "获取用户屏蔽聊天列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户唯一标识",
                        "name": "metaId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.UserBlockedChats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/v1/push/get_user_token": {
            "get": {
                "description": "根据用户 metaId 获取该用户的所有推送令牌",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push API"
                ],
                "summary": "根据 metaId 获取用户推送令牌",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户唯一标识",
                        "name": "metaId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.UserPushTokens"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/v1/push/get_user_tokens_list": {
            "get": {
                "description": "分页获取所有用户的推送令牌列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push API"
                ],
                "summary": "获取用户推送令牌列表（分页）",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认为1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页大小，默认为10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pebble_service.PaginatedUserTokens"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/v1/push/remove_blocked_chat": {
            "post": {
                "description": "移除用户对某个群聊或私聊的屏蔽",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push API"
                ],
                "summary": "移除屏蔽聊天",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RemoveBlockedChatReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/v1/push/remove_user_all_tokens": {
            "post": {
                "description": "移除指定用户的所有推送令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push API"
                ],
                "summary": "移除用户所有推送令牌",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RemoveUserAllTokensReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/v1/push/remove_user_token": {
            "post": {
                "description": "移除指定用户在指定平台的推送令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push API"
                ],
                "summary": "移除用户推送令牌",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RemoveUserTokenReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        },
        "/v1/push/set_user_tokens": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "为指定用户在指定平台设置推送令牌，支持Token唯一性检查。Token本身就是设备的唯一标识，如果Token已被其他用户使用，会自动从原用户中移除该平台的令牌。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push API"
                ],
                "summary": "设置用户推送令牌",
                "parameters": [
                    {
                        "description": "请求参数（metaId、platform、token）",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SetUserTokensReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/respond.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BlockedChat": {
            "type": "object",
            "required": [
                "chatId",
                "userId"
            ],
            "properties": {
                "blockedAt": {
                    "description": "屏蔽时间",
                    "type": "integer"
                },
                "chatId": {
                    "description": "群ID或私聊ID",
                    "type": "string"
                },
                "chatType": {
                    "description": "聊天类型 (group, private)",
                    "type": "string"
                },
                "reason": {
                    "description": "屏蔽原因",
                    "type": "string"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "string"
                }
            }
        },
        "models.UserBlockedChats": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "blockedChats": {
                    "description": "屏蔽的聊天列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BlockedChat"
                    }
                },
                "updatedAt": {
                    "description": "最后更新时间",
                    "type": "integer"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "string"
                }
            }
        },
        "models.UserPushTokens": {
            "type": "object",
            "required": [
                "metaId"
            ],
            "properties": {
                "metaId": {
                    "description": "用户唯一标识",
                    "type": "string"
                },
                "tokens": {
                    "description": "平台-\u003e令牌映射 {\"expo\": \"ExponentPushToken[...]\", \"fcm\": \"fcm_token_123\"}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "description": "最后更新时间",
                    "type": "integer"
                }
            }
        },
        "pebble_service.PaginatedUserTokens": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "description": "是否有下一页",
                    "type": "boolean"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total": {
                    "description": "总数量",
                    "type": "integer"
                },
                "totalPages": {
                    "description": "总页数",
                    "type": "integer"
                },
                "users": {
                    "description": "用户令牌列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserPushTokens"
                    }
                }
            }
        },
        "request.AddBlockedChatReq": {
            "type": "object",
            "required": [
                "chatId",
                "chatType",
                "metaId"
            ],
            "properties": {
                "chatId": {
                    "type": "string"
                },
                "chatType": {
                    "description": "聊天类型：group, private",
                    "type": "string"
                },
                "metaId": {
                    "type": "string"
                },
                "reason": {
                    "description": "屏蔽原因（可选）",
                    "type": "string"
                }
            }
        },
        "request.RemoveBlockedChatReq": {
            "type": "object",
            "required": [
                "chatId",
                "metaId"
            ],
            "properties": {
                "chatId": {
                    "type": "string"
                },
                "metaId": {
                    "type": "string"
                }
            }
        },
        "request.RemoveUserAllTokensReq": {
            "type": "object",
            "required": [
                "metaId"
            ],
            "properties": {
                "metaId": {
                    "type": "string"
                }
            }
        },
        "request.RemoveUserTokenReq": {
            "type": "object",
            "required": [
                "metaId",
                "platform"
            ],
            "properties": {
                "metaId": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                }
            }
        },
        "request.SetUserTokensReq": {
            "type": "object",
            "required": [
                "metaId",
                "platform",
                "token"
            ],
            "properties": {
                "metaId": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "token": {
                    "description": "Token本身就是设备的唯一标识",
                    "type": "string"
                }
            }
        },
        "respond.Response": {
            "description": "统一的 API 响应格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "processingTime": {
                    "type": "integer",
                    "example": 123
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    }
}